interface: Introduce lightweight crate for instruction + id
#### Problem

In order to publish the v3 SDK crates and have them usable in Agave, we
also need to have SPL crates using the v3 SDK crates. However, we have a
circular dependency between Agave and SPL which currently makes this
impossible.

The overall plan is to have Agave only use "interface" crates from SPL,
which have no dependencies on Agave crates.

You can see more info about the project at https://github.com/orgs/anza-xyz/projects/27

Instruction-padding is pretty simple, since it just exposes an
instruction creator and id to Agave.

#### Summary of changes

Create the new interface crate with the instruction and id.

The Rust scripts needed to be adapted to having more than one crate, so
I copied all the scripts from token-2022 and updated the commands in
package.json accordingly.
