+++ interface/Cargo.toml
@@ -1,35 +1,25 @@
 [package]
-name = "spl-instruction-padding"
-version = "0.3.0"
-description = "Solana Program Library Instruction Padding Program"
+name = "spl-instruction-padding-interface"
+version = "0.1.0"
+description = "Solana Program Library Instruction Padding Interface"
 authors = ["Anza Maintainers <maintainers@anza.xyz>"]
 repository = "https://github.com/solana-program/instruction-padding"
 license = "Apache-2.0"
 homepage = "https://solana.com/"
 edition = "2021"
 
-[features]
-no-entrypoint = []
-test-sbf = []
-
 [dependencies]
-solana-account-info = "2.3.0"
-solana-cpi = "2.2.1"
+num_enum = "0.7.4"
 solana-instruction = { version = "2.2.1", features = ["std"] }
-solana-program-entrypoint = "2.3.0"
 solana-program-error = "2.2.2"
 solana-pubkey = "2.2.1"
-spl-instruction-padding-interface = { path = "../interface" }
 
 [dev-dependencies]
 solana-program = "2.3.0"
-solana-program-test = "2.3.6"
-solana-sdk = "2.2.1"
-solana-system-interface = "1"
 static_assertions = "1.1.0"
 
 [lib]
-crate-type = ["cdylib", "lib"]
+crate-type = ["lib"]
 
 [package.metadata.docs.rs]
 targets = ["x86_64-unknown-linux-gnu"]
