+++ interface/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod instruction;
+
+solana_pubkey::declare_id!("iXpADd6AW1k5FaaXum5qHbSqyd7TtoN6AD7suVa83MF");
