+++ program/src/lib.rs
@@ -2,8 +2,8 @@ mod entrypoint;
 pub mod instruction;
 pub mod processor;
 
+pub use spl_instruction_padding_interface::{check_id, id, ID};
 pub use {
     solana_account_info, solana_cpi, solana_instruction, solana_program_entrypoint,
     solana_program_error, solana_pubkey,
 };
-solana_pubkey::declare_id!("iXpADd6AW1k5FaaXum5qHbSqyd7TtoN6AD7suVa83MF");
