+++ scripts/rust/audit.mjs
@@ -20,6 +20,17 @@ const advisories = [
   // URL:       https://rustsec.org/advisories/RUSTSEC-2024-0376
   // Solution:  Upgrade to >=0.12.3
   'RUSTSEC-2024-0376',
+
+  // Crate:     idna
+  // Version:   0.1.5
+  // Title:     `idna` accepts Punycode labels that do not produce any non-ASCII when decoded
+  // Date:      2024-12-09
+  // ID:        RUSTSEC-2024-0421
+  // URL:       https://rustsec.org/advisories/RUSTSEC-2024-0421
+  // Solution:  Upgrade to >=1.0.0
+  // need to solve this dependency tree:
+  // jsonrpc-core-client v18.0.0 -> jsonrpc-client-transports v18.0.0 -> url v1.7.2 -> idna v0.1.5
+  'RUSTSEC-2024-0421',
 ];
 const ignores = []
 advisories.forEach(x => {
