+++ scripts/rust/lint.mjs
@@ -2,11 +2,12 @@
 import 'zx/globals';
 import {
   cliArguments,
-  getProgramFolders,
   getToolchainArgument,
   popArgument,
   workingDirectory,
 } from '../utils.mjs';
+
+const [folder, ...args] = cliArguments();
 
 // Configure arguments here.
 const lintArgs = [
@@ -16,19 +17,16 @@ const lintArgs = [
   '--',
   '--deny=warnings',
   '--deny=clippy::arithmetic_side_effects',
-  ...cliArguments(),
+  ...args,
 ];
 
 const fix = popArgument(lintArgs, '--fix');
 const toolchain = getToolchainArgument('lint');
 
-// Lint the programs using clippy.
-for (const folder of getProgramFolders()) {
-  const manifestPath = path.join(workingDirectory, folder, 'Cargo.toml');
+const manifestPath = path.join(workingDirectory, folder, 'Cargo.toml');
 
-  if (fix) {
-    await $`cargo ${toolchain} clippy --manifest-path ${manifestPath} --fix ${lintArgs}`;
-  } else {
-    await $`cargo ${toolchain} clippy --manifest-path ${manifestPath} ${lintArgs}`;
-  }
+if (fix) {
+  await $`cargo ${toolchain} clippy --manifest-path ${manifestPath} --fix ${lintArgs}`;
+} else {
+  await $`cargo ${toolchain} clippy --manifest-path ${manifestPath} ${lintArgs}`;
 }
